+++ setup/DeepseekHarness.Setup/Views/InstallWindow.cs
@@ -152,7 +152,7 @@ public class InstallWindow : WindowBase
         }
     }
 
-    private static string AppVersion => Assembly.GetExecutingAssembly().GetName().Version?.ToString(3) ?? "1.0.3";
+    private static string AppVersion => Assembly.GetExecutingAssembly().GetName().Version?.ToString(3) ?? "1.1.0";
 
     private void SetDefaultInstallFolder()
     {